+++ packages/eslint-config/rules/base/es6.js
@@ -71,21 +71,16 @@ module.exports = {
     'prefer-spread': 'error',
     'prefer-template': 'error',
     'require-yield': 'error',
-    'sort-imports': 'off', // Rule not refined enough for our liking
-    /*
-     * Here is the closest approximation of what we'd like to achieve with the available options:
-     *
-     * 'sort-imports': [
-     *   'error',
-     *   {
-     *     ignoreCase: false,
-     *     ignoreDeclarationSort: true,
-     *     ignoreMemberSort: false,
-     *     memberSyntaxSortOrder: ['single', 'multiple', 'all', 'none'],
-     *   },
-     * ],
-     *
-     */
+    'sort-imports': [
+      'warn',
+      {
+        allowSeparatedGroups: true,
+        ignoreCase: true,
+        ignoreDeclarationSort: false,
+        ignoreMemberSort: false,
+        memberSyntaxSortOrder: ['none', 'all', 'single', 'multiple'],
+      },
+    ],
     'symbol-description': 'error',
   },
 };