+++ packages/eslint-config/rules/deprecated/deprecated.js
@@ -28,10 +28,10 @@ module.exports = {
     'require-jsdoc': 'off', // ESLint - v5.10.0
     'valid-jsdoc': 'off', // ESLint - v5.10.0
     'no-catch-shadow': 'off', // ESLint - v5.1.0 - no-shadow
+    'indent-legacy': 'off', // ESLint - v4.0.0 - indent
     'lines-around-directive': 'off', // ESLint - v4.0.0 - padding-line-between-statements
     'newline-after-var': 'off', // ESLint - v4.0.0 - padding-line-between-statements
     'newline-before-return': 'off', // ESLint - v4.0.0 - padding-line-between-statements
-    'indent-legacy': 'off', // ESLint - v4.0.0 - indent
     'prefer-reflect': 'off', // ESLint - v3.9.0
     'no-native-reassign': 'off', // ESLint - v3.3.0 - no-global-assign
     'no-negated-in-lhs': 'off', // ESLint - v3.3.0 - no-unsafe-negation